feat: 엔티티 관계 설정
스터디룸과 채팅메시지의 연관 관계 설정
