+++ techere/src/main/java/com/techere/project/TechereApplication.java
@@ -2,7 +2,9 @@ package com.techere.project;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
 
+@EnableJpaAuditing
 @SpringBootApplication
 public class TechereApplication {
