+++ techere/src/main/java/com/techere/project/domain/kafkachat/dto/mapper/ChatMapper.java
@@ -1,8 +1,7 @@
-package com.techere.project.kafkachat.dto.mapper;
+package com.techere.project.domain.kafkachat.dto.mapper;
 
-import com.techere.project.kafkachat.dto.request.MessageRequestDto;
-import com.techere.project.kafkachat.model.KafkaMessage;
-import lombok.RequiredArgsConstructor;
+import com.techere.project.domain.kafkachat.dto.request.MessageRequestDto;
+import com.techere.project.domain.kafkachat.model.KafkaMessage;
 import org.springframework.stereotype.Component;
 
 import java.time.LocalDateTime;
