+++ techere/src/main/java/com/techere/project/domain/kafkachat/repository/ChatRepository.java
@@ -1,0 +1,8 @@
+package com.techere.project.domain.kafkachat.repository;
+
+import com.techere.project.domain.kafkachat.model.Message;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface ChatRepository extends JpaRepository<Message, Long> {
+
+}
