--- techere/src/main/java/com/techere/project/kafkachat/repository/ChatRepository.java
@@ -1,9 +1,0 @@
-package com.techere.project.kafkachat.repository;
-
-import com.techere.project.kafkachat.model.KafkaMessage;
-import com.techere.project.kafkachat.model.Message;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface ChatRepository extends JpaRepository<Message, Long> {
-
-}
