+++ techere/src/main/resources/application.properties
@@ -7,6 +7,3 @@ spring.datasource.password=${SPRING_DATA_SOURCE_PASSWORD}
 spring.main.allow-bean-definition-overriding=true
 spring.jpa.hibernate.ddl-auto=update
 spring.mvc.hiddenmethod.filter.enabled=true
-
-spring.kafka.bootstrap-servers=localhost:9092
-spring.kafka.consumer.group-id=test
